refactor: concern-split AI planning near-cap libs and ratchet to 300 (#4365)
Slice A: split fourteen planning modules at 280–316 physical lines into
sibling libraries (no part-of), keep public barrels, and lower
repo.ai_source_file_size from 350 to 300 with an empty grandfather.
