+++ SPEC/program/repo-lint.md
@@ -189,7 +189,7 @@
 | `tool/check_ai_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_ai/lib/**` so planning/perception sub-files stay proper libraries with explicit imports (mirrors economy/world/turn/models no-part gates; Refs #4079); rule `repo.ai_no_part_directives`. Also hosts the shared parameterized scanner used by `repo.ai_contracts_no_part_directives` (Refs #4084). |
 | `tool/check_ai_contracts_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_ai_contracts/lib/**` so Full AI civilian selection / heuristics / localization sub-files stay proper libraries with explicit imports (mirrors `repo.ai_no_part_directives`; Refs #4084) — see `SPEC/program/ai-contracts-no-part-directives.md`; rule `repo.ai_contracts_no_part_directives` |
 | `tool/check_ai_contracts_h8_test_shared_fixtures.dart` | Forbid local `_flaggedSellerGame` redeclarations in the three H8 seller feedstock acquisition-target pins and local `_supplierGame` redeclarations in the OW feedstock prospect localization / mineral-eligibility pins under `packages/colonizethis_ai_contracts/test/`. Pins MUST import `flaggedSellerGame` / `supplierGame` from `test/support/h8_flagged_seller_game.dart` and `test/support/h8_supplier_prospect_game.dart`. The gate also fails when either shared support file is missing (Refs #4084); rule `repo.ai_contracts_h8_test_shared_fixtures` |
-| `tool/check_ai_source_file_size.dart` | Enforce `packages/colonizethis_ai/lib/src/**` Dart files at **350 physical lines or fewer**, with an optional shrink-only grandfather allowlist that fails when entries are missing or now under-cap (mirrors save/data size gates; headroom ratchet after #4104 Slice B, #4239 Slice A, and #4310 Slice B near-cap splits; Refs #4079 Slice C, #4104 Slice B, #4239 Slice A, #4310) — see `SPEC/program/ai-source-file-size.md`; rule `repo.ai_source_file_size` |
+| `tool/check_ai_source_file_size.dart` | Enforce `packages/colonizethis_ai/lib/src/**` Dart files at **300 physical lines or fewer**, with an optional shrink-only grandfather allowlist that fails when entries are missing or now under-cap (mirrors save/data size gates; headroom ratchet after #4104 Slice B, #4239 Slice A, #4310 Slice B, and #4365 Slice A near-cap splits; Refs #4079 Slice C, #4104 Slice B, #4239 Slice A, #4310, #4365) — see `SPEC/program/ai-source-file-size.md`; rule `repo.ai_source_file_size` |
 | `tool/check_models_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_models/test/**` exceeds 400 physical lines (aligns with diplomacy/economy test file caps; Refs #4068 Slice D); rule `repo.models_test_file_size` |
 | `tool/check_models_test_mirrors_lib.dart` | Keep the `packages/colonizethis_models/test/**` tree mirrored to `lib/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root is a violation — tests must live in subdirectories such as `test/src/`, `test/app_events/`, or `test/diplomacy/` that mirror `lib/src/` (Refs #4334); rule `repo.models_test_mirrors_lib` |
 | `tool/check_world_topology_identity_home.dart` | Require topology identity helpers (`indexTopologyNodesByRegion`, `provinceTopologyNodeId`, `provinceIdsAdjacentToSeaZone`, `regionIdForSeaZone`) to be defined in `packages/colonizethis_world/lib/src/world/topology_identity.dart` and forbid re-declaring them as top-level functions in `naval.dart` (re-export allowed; Refs #3968); rule `repo.world_topology_identity_home` |
@@ -594,9 +594,9 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a gated H8 seller acquisition-target pin that redeclares `Game _flaggedSellerGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the run fails listing the offending file and exits `1` (Refs #4084).
 - Given a gated OW prospect localization or mineral-eligibility pin that redeclares `Game _supplierGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the run fails listing the offending file and exits `1` (Refs #4084).
 - Given either shared support file `h8_flagged_seller_game.dart` or `h8_supplier_prospect_game.dart` is missing under `packages/colonizethis_ai_contracts/test/support/`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the rule exits `1` (Refs #4084).
-- Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is at or below 350 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the rule passes without violations and exits `0` (Refs #4079, #4104, #4239, #4310).
-- Given a hand-written `packages/colonizethis_ai/lib/src/**` Dart file with more than 350 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the run fails naming that file and exits `1` (Refs #4079, #4104, #4239, #4310).
-- Given a grandfather allowlist entry for `repo.ai_source_file_size` that names a missing file or a file now at or below 350 physical lines, when the rule runs, then the run fails reporting a stale grandfather entry and exits `1` (Refs #4079, #4104, #4239, #4310).
+- Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is at or below 300 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the rule passes without violations and exits `0` (Refs #4079, #4104, #4239, #4310, #4365).
+- Given a hand-written `packages/colonizethis_ai/lib/src/**` Dart file with more than 300 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the run fails naming that file and exits `1` (Refs #4079, #4104, #4239, #4310, #4365).
+- Given a grandfather allowlist entry for `repo.ai_source_file_size` that names a missing file or a file now at or below 300 physical lines, when the rule runs, then the run fails reporting a stale grandfather entry and exits `1` (Refs #4079, #4104, #4239, #4310, #4365).
 - Given every `.dart` file under `packages/colonizethis_models/test/**` is at or below 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the rule passes without violations and exits `0` (Refs #4068).
 - Given a `packages/colonizethis_models/test/**` Dart file with more than 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the run fails naming that file and exits `1` (Refs #4068).
 - Given every models test Dart source lives in a subdirectory under `packages/colonizethis_models/test/` (no loose files at the `test/` root except paths under `test/support/`), when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_mirrors_lib`, then the rule passes and exits `0` (Refs #4334).
